requirements.txt
langgraph
langchain
qdrant_client

aiogram~=3.0.0b8
Jinja2~=3.1.3
typing~=3.7.4.3
python-dotenv~=1.0.1